requirements.txt
eupy==1.0.23
matplotlib==3.10.3
numpy==2.2.5
pandas==2.2.3
peft==0.15.2
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
rich==14.0.0
torch==2.5.1
tqdm==4.67.1
tqdm==4.66.5
transformers==4.47.1